% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsupkg_next_version.R
\name{confirm_git_branch_is_as_expected}
\alias{confirm_git_branch_is_as_expected}
\title{confirm_git_branch_is_as_expected}
\usage{
confirm_git_branch_is_as_expected(branch_expected, directory_to_check,
  from_where = c("system", "System", "R_options"), fail.if.not = TRUE,
  showWarnings = !fail.if.not)
}
\arguments{
\item{from_where}{enum string of length 1. Where are we checking the git branch? Are we using the system command, or are we checking what branch we have saved in R_options?}
}
\description{
checks that the system git branch matches an exepected git branch
}
