% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_source_package.R
\name{.refresh_rsu_pkg}
\alias{.refresh_rsu_pkg}
\title{.refresh_rsu_pkg}
\usage{
.refresh_rsu_pkg(
  pkg,
  parent_folder = .get_rsu_homeDir(),
  git_pull = FALSE,
  document_too = FALSE,
  diff_shown.for_documenting = verbose,
  source_instead_of_devtools = FALSE,
  verbose = TRUE
)
}
\description{
This is a wrapper to a wrapper! hahah. calls .rsu_source_package_files()
}
\details{
Because I have gotten so use to .refresh_general() and similar, I now need a .refresh_rsu_pkg()

Although truthfully, I should probably just take a moment to type one for each, and the end.

## TODO:  The other .refresh_.. should probably call this, no?
}
