% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_source_package.R
\name{.rsu_document_package_files}
\alias{.rsu_document_package_files}
\title{.rsu_document_package_files}
\usage{
.rsu_document_package_files(
  pkg = .rsu_pkgs_strings(only_installed = only_installed),
  folder = file.path(parent_folder, pkg, "R"),
  parent_folder = .get_rsu_homeDir(),
  git_pull = FALSE,
  only_installed = TRUE,
  diff_shown = TRUE,
  verbose = TRUE,
  verbose.devtools = verbose,
  verbose.git_cmd = verbose,
  verbose.diff = diff_shown && verbose
)
}
\description{
.rsu_document_package_files
}
