% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_source_package.R
\name{.rsu_source_package_files}
\alias{.rsu_source_package_files}
\title{.rsu_source_package_files}
\usage{
.rsu_source_package_files(
  pkg = .rsu_pkgs_strings(only_installed = only_installed),
  folder = file.path(parent_folder, pkg, "R"),
  parent_folder = .get_rsu_homeDir(),
  git_pull = FALSE,
  only_installed = TRUE,
  document_too = FALSE,
  diff_shown.for_documenting = verbose,
  source_instead_of_devtools = FALSE,
  verbose = TRUE
)
}
\description{
.rsu_source_package_files
}
