% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell6s.R
\name{rm_in_file}
\alias{rm_in_file}
\alias{run_6s}
\alias{run_simulate_aod}
\title{run_simulate_aod}
\usage{
run_6s(
  in_file = "in.txt",
  exe_file = "6s.exe",
  out_file = "out.txt",
  rm_out_file = TRUE,
  rm_in_file = TRUE
)

run_simulate_aod(
  band_file = system.file("OLCI_BN15.csv", package = "rsutlis"),
  in_ref = system.file("in_ref.txt", package = "rsutlis"),
  aod550_arr = seq(0.1, 0.5, 0.1),
  verbose = TRUE,
  sleep = TRUE,
  ...
)
}
\arguments{
\item{in_file}{in_file}

\item{exe_file}{6s.exe file}

\item{out_file}{out.txt file generated from 6s.exe}

\item{rm_out_file}{Remove the out.txt after runing 6s. Default is \code{TRUE}.}

\item{rm_in_file}{Remove the in.txt after running 6s. Default is \code{TRUE}.}

\item{band_file}{band_file}

\item{in_ref}{in.txt reference}

\item{aod550_arr}{simulated aod550}

\item{verbose}{TRUE}

\item{sleep}{TRUE}

\item{...}{parameters of the function \code{run_6s}}
}
\value{
None

A list including the output reflectance.
}
\description{
This function shells 6s.exe in the current work path.

Simulate the reflectance of aerosol varying on AOD550.
}
\examples{
\dontrun{
result <- run_simulate_aod(aod550_arr = seq(0.1, 0.5, 0.1))
ref_aerosols <- result$spec_aerosols
tmp <- reshape2::melt(ref_aerosols, id=c("aod550"))
if(require(ggplot2)) {
tmp$variable <- as.numeric(as.vector(tmp$variable))
ggplot(data = tmp) + geom_path(aes(x = variable, y = value / pi,
group = aod550, color = aod550), size=1) + scale_color_viridis_c()
}
}

}
