% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_over_rasters.R
\name{points_over_rasters}
\alias{points_over_rasters}
\title{points_over_rasters}
\usage{
points_over_rasters(im, pts, show.map = FALSE)
}
\arguments{
\item{im}{Raster object or file path could be read}

\item{pts}{data.frame of points with name, lon, and lat}

\item{show.map}{Whether to plot the position and raster frame on \link{leaflet}.
Default is \code{FALSE}.}
}
\value{
Data frame extracted from the raster file
}
\description{
points_over_rasters
}
\note{
The crs of \code{pts} should be \code{+proj=longlat +datum=WGS84 +no_defs}.
}
\examples{
library(raster)
library(rsutlis)
f <- system.file("external/test.grd", package="raster")
im <- raster(f)
pts <- data.frame(name = 1:3,
                  lon = c(5.73576, 5.73747, 5.74089),
                  lat = c(50.97330, 50.96790, 50.97942))
result <- points_over_rasters(im, pts)

}
